NATIVE-193: Rename file and make ErrorCodes default export
