+++ src/modules/endpoint/reducers/__tests__/citiesReducer.spec.js
@@ -3,7 +3,7 @@
 import type { CitiesStateType } from '../../../app/StateType'
 import citiesReducer from '../citiesReducer'
 import { CityModel } from '@integreat-app/integreat-api-client'
-import { ErrorCodes } from '../../../error/ErrorCode'
+import ErrorCodes from '../../../error/ErrorCodes'
 
 describe('citiesReducer', () => {
   it('should set status to loading on FETCH_CITIES', () => {
