+++ src/modules/endpoint/reducers/__tests__/cityContentReducer.spec.js
@@ -4,7 +4,7 @@ import type { CityContentActionType, FetchCategoryActionType, FetchEventActionTy
 import { CategoriesMapModel, LanguageModel } from '@integreat-app/integreat-api-client'
 import cityContentReducer from '../cityContentReducer'
 import type { CityContentStateType } from '../../../app/StateType'
-import { ErrorCodes } from '../../../error/ErrorCode'
+import ErrorCodes from '../../../error/ErrorCodes'
 
 describe('cityContentReducer', () => {
   const switchContentLanguageAction = {
