+++ src/modules/error/hocs/withPayloadProvider.js
@@ -10,7 +10,7 @@ import { type Dispatch } from 'redux'
 import { wrapDisplayName } from 'recompose'
 import FailureContainer from '../containers/FailureContainer'
 import { LOADING_TIMEOUT } from '../../common/constants'
-import type { ErrorCodeType } from '../ErrorCode'
+import type { ErrorCodeType } from '../ErrorCodes'
 
 export type RouteNotInitializedType = {| status: 'routeNotInitialized' |}
 export type LoadingType = {| status: 'loading' |}
